feat: 당첨 번호 로직 - 1~45 범위 검증
사용자에게 입력 받은 당첨 번호가 1~45 이내에 속한 숫자인지 검증합니다.
위 범위 외의 숫자인 경우 IllegalArgumentException을 던지도록 합니다.

+++ src/main/java/lotto/model/Winning.java
@@ -1,12 +1,13 @@
 package lotto.model;
 
+import java.math.BigInteger;
 import java.util.List;
 
 public class Winning {
-    List<Long> winNumbers;
-    Long bonusNumber;
+    List<BigInteger> winNumbers;
+    BigInteger bonusNumber;
 
-    public Winning(List<Long> winNumbers, Long bonusNumber) {
+    public Winning(List<BigInteger> winNumbers, BigInteger bonusNumber) {
         this.winNumbers = winNumbers;
         this.bonusNumber = bonusNumber;
     }
